- Refactor to improve code
