+++ features/backup/src/main/java/ir/mehdiyari/krypt/backup/DataRoute.kt
@@ -1,19 +1,26 @@
 package ir.mehdiyari.krypt.backup
 
 import android.annotation.SuppressLint
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
 import android.widget.Toast
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
+import androidx.documentfile.provider.DocumentFile
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import ir.mehdiyari.krypt.core.designsystem.theme.KryptTheme
@@ -28,14 +35,33 @@ internal fun DataRoute(
     modifier: Modifier,
 ) {
     val managerStoragePermissionState = remember { mutableStateOf(false) }
-    val coroutineScope = rememberCoroutineScope()
 
-    val showDirectoryChooser = remember { mutableStateOf(false) }
-    val snackbarHostState = remember { SnackbarHostState() }
+    val chooseDirectorySnackbarMsg =
+        stringResource(id = R.string.choose_backup_directory_description)
+    val context = LocalContext.current
+
+    var selectedDirectory: DocumentFile? by remember { mutableStateOf(null) }
+    val launcher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.StartActivityForResult(),
+        onResult = { result ->
+            if (result.resultCode == Activity.RESULT_OK) {
+                val uri: Uri? = result.data?.data
+                uri?.let {
+                    selectedDirectory = DocumentFile.fromTreeUri(
+                        context, it
+                    )
+                }
+            }
+        }
+    )
+
+    selectedDirectory?.let {
+        viewModel.backupNow(it.uri)
+    }
+
     DataScreenScaffold(
         modifier = modifier,
-        onNavigationClicked = onNavigationClicked,
-        snackbarHostState = snackbarHostState
+        onNavigationClicked = onNavigationClicked
     ) {
         Column(
             modifier = modifier
@@ -47,20 +73,16 @@ internal fun DataRoute(
             val lastBackupState = viewModel.lastBackupDateTime.collectAsStateWithLifecycle()
             val backupState = viewModel.backupViewState.collectAsStateWithLifecycle()
             BackupView(modifier, lastBackupState, backupState, backupNowClick = {
-                showDirectoryChooser.value = true
-            })
-            if (showDirectoryChooser.value) {
                 if (!checkIfAppIsStorageManager()) {
                     managerStoragePermissionState.value = true
                 } else {
-                    ChooseDirectoryView(
-                        coroutineScope = coroutineScope,
-                        snackbarHostState = snackbarHostState,
-                        onSelectDirectory = viewModel::backupNow
-                    )
+                    Toast.makeText(
+                        context, chooseDirectorySnackbarMsg, Toast.LENGTH_LONG
+                    ).show()
+                    val intent = Intent(Intent.ACTION_OPEN_DOCUMENT_TREE)
+                    launcher.launch(intent)
                 }
-
-            }
+            })
 
             val backupList = viewModel.backups.collectAsStateWithLifecycle()
 
@@ -90,7 +112,6 @@ internal fun DataRoute(
         }
     }
 
-    val context = LocalContext.current
     if (managerStoragePermissionState.value) {
         ManageExternalPermissionDialog(modifier = modifier, state = managerStoragePermissionState) {
             managerStoragePermissionState.value = false
@@ -109,9 +130,7 @@ internal fun DataRoutePreview(
     ) backupList: List<BackupViewData>
 ) {
     KryptTheme {
-        DataScreenScaffold(modifier = Modifier, snackbarHostState = remember {
-            SnackbarHostState()
-        }) {
+        DataScreenScaffold(modifier = Modifier) {
             Column {
                 FileSizePreview()
                 BackupViewPreview()
